+++ testcases/fail/style_indent.sv
@@ -1,0 +1,10 @@
+module M;
+ if (a)
+   a = 0;
+   else
+     a = 1;
+   // comment
+/*
+ comment
+   */
+endmodule