+++ testcases/fail/style_keyword_0or1space.sv
@@ -1,0 +1,7 @@
+module A;
+  function foo();
+    if (a)
+      return  ; // multiple spaces after `return`.
+  endfunction
+endmodule
+