+++ testcases/fail/style_keyword_1or2space.sv
@@ -1,0 +1,5 @@
+module M (
+  input   a,
+  inout   b   // multiple spaces after `input` or `inout` keywords
+);
+endmodule