+++ testcases/fail/style_keyword_newline.sv
@@ -1,0 +1,8 @@
+module A;
+  generate
+    case (foo)
+      123: a = b;
+    endcase if (foo) a = b; // no newline after `endcase`
+  endgenerate   // multiple spaces then comment after `endgenerate`
+endmodule
+